refactor(topic: separate hook for topic
