+++ src/app/[locale]/topics/[topicId]/(topic)/context/TopicWorkspaceContext.tsx
@@ -15,6 +15,7 @@ import React, {
 import { isAfter } from 'date-fns';
 import { IAnkiSetting } from '@/types/anki-setting/ankiSetting.type';
 import { TopicWorkspaceTabValue } from '../types';
+import useFlashCardWorkSpace from '../hooks/useFlashCardWorkSpace';
 
 export type TypeTopicId = number;
 interface ContextType {
@@ -55,11 +56,6 @@ export function TopicWorkspaceProvider({ children, topicIdInit }: IProviderProps
     const [tab, setTab] = useState<TopicWorkspaceTabValue>(DEFAULT_TAB);
 
     const [topic, setTopic] = useState<ITopic | null>(null);
-    const [flashcards, setFlashcards] = useState<IFlashcard[] | null>(null);
-    const [learningFlashcards, setLearningFlashcards] = useState<IDueAnkiCard[] | null>(null);
-    const [ankiSettings, setAnkiSettings] = useState<{ settings: IAnkiSetting[]; activeSettingId: number } | null>(
-        null,
-    );
 
     const topicIdRef = useRef<TypeTopicId>(topicIdInit);
 
@@ -67,46 +63,19 @@ export function TopicWorkspaceProvider({ children, topicIdInit }: IProviderProps
 
     const [isPdfViewerFullscreen, setIsPdfViewerFullScreen] = useState<boolean>(false);
 
+    const {
+        flashcards,
+        learningFlashcards,
+        ankiSettings,
+        setFlashcards,
+        setLearningFlashcards,
+        setAnkiSettings,
+        onReviewCard,
+    } = useFlashCardWorkSpace();
+
     const setTopicId = useCallback((topicIdArg: TypeTopicId) => {
         topicIdRef.current = topicIdArg;
     }, []);
-
-    const onReviewCard = useCallback(
-        ({ currentCard, reviewedCard }: { currentCard: IDueAnkiCard; reviewedCard: IAnkiCardReviewed | null }) => {
-            const updatedLearningFlashcards = [...(learningFlashcards ?? [])];
-            updatedLearningFlashcards.shift();
-            if (reviewedCard) {
-                // INSERT this card to a suitable position (to maintain ORDER by nextReview)
-                let inserted = false;
-                for (let i = 0; i < updatedLearningFlashcards.length; ++i) {
-                    const card = updatedLearningFlashcards[i];
-                    if (isAfter(reviewedCard.nextReview, card.nextReview)) continue;
-                    updatedLearningFlashcards.splice(i, 0, {
-                        ...currentCard,
-                        nextReview: reviewedCard.nextReview,
-                        status: reviewedCard.status,
-                        nextReviewDataByRatings: reviewedCard.nextReviewDataByRatings,
-                    });
-                    inserted = true;
-                    break;
-                }
-
-                if (!inserted) {
-                    updatedLearningFlashcards.push({
-                        ...currentCard,
-                        nextReview: reviewedCard.nextReview,
-                        status: reviewedCard.status,
-                        nextReviewDataByRatings: reviewedCard.nextReviewDataByRatings,
-                    });
-                    inserted = true;
-                }
-            }
-
-            setLearningFlashcards(updatedLearningFlashcards);
-            return updatedLearningFlashcards;
-        },
-        [learningFlashcards],
-    );
 
     const value = useMemo(
         () => ({
